Implement loop to read one double at a time, track smallest and largest values, and provide user feedback on input. part 6 of the drill done!

+++ Section 3 computation/Drills/part3_and_4.cpp
@@ -6,11 +6,13 @@
 // the larger value is: followed by the larger value.
 // Learning c++ 
 
+// Change the program so that it uses `double`s instead of `int`s.
+
 #include <iostream>
 
 int main() {
-  int val1;
-  int val2;
+  double val1;
+  double val2;
 
   std::string stop;
